requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2

# Graph database
neo4j>=4.4.0
py2neo>=2021.2.3

# Machine Learning
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.10.0

# Data processing and utilities
python-dateutil>=2.8.2
pytz>=2021.3
tqdm>=4.62.3

# Testing and development
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910

# Security and compliance
cryptography>=3.4.7
python-jose>=3.3.0
passlib>=1.7.4 